Revamped the UI using fluent_ui and bitsdojo_window packages
