+++ lib/about.dart
@@ -1,5 +1,5 @@
+import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter/gestures.dart';
-import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher_string.dart';
 
 class AboutApp extends StatelessWidget {
@@ -14,8 +14,9 @@ class AboutApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Dialog(
-      child: Container(
+    return ContentDialog(
+      constraints: const BoxConstraints(maxWidth: double.infinity),
+      content: Container(
         padding: const EdgeInsets.all(24),
         constraints: BoxConstraints(
             maxWidth: MediaQuery.of(context).size.width * 0.6,
