+++ pubspec.yaml
@@ -27,6 +27,8 @@ dependencies:
   package_info_plus: ^4.2.0
   shared_preferences: ^2.2.2
   version: ^3.0.2
+  fluent_ui: ^4.8.7
+  bitsdojo_window: ^0.1.6
 
 dev_dependencies:
   flutter_test:
